Rough Angle Calculations V1
This is the initial script testing for the angle calculations.

It incorporates a new directory for functionality scripts, with one current script that can be used to determine the HSV value of an object using two live feeds (One showing HSV, One showing camera feed) and an HSV slider.

The second directory is for in-progress testing scripts and includes one script that is very robust and inconsistently calculates the angle of an object using an arbitrary focal point.

Both scripts still need to have the comments updated and the second version will have new angle calculations.